+++ src/QNEthernet.cpp
@@ -459,8 +459,7 @@ bool EthernetClass::setMACAddressAllowed(const uint8_t mac[6],
     return false;
   }
 #ifndef QNETHERNET_ENABLE_PROMISCUOUS_MODE
-  enet_set_mac_address_allowed(mac, flag);
-  return true;
+  return enet_set_mac_address_allowed(mac, flag);
 #else
   return flag;  // Can't disallow MAC addresses
 #endif  // !QNETHERNET_ENABLE_PROMISCUOUS_MODE